Viewplane
Got the viewplane done

+++ RayTracer/RayTracer/Header/definitions.h
@@ -4,4 +4,4 @@
 typedef glm::vec3 vec3;
 typedef glm::dvec3 dvec3;
 typedef glm::dvec3 colorDbl; //color vector
-typedef glm::dvec3 direction; //direction vector
+typedef glm::vec3 direction; //direction vector
